Add files via upload
Added an alternative test script that allows UMX to process a target dataset without performing evaluation. This script is based on the original test.py file and the same arguments can therefore also be parsed as in the original. However, the default arguments in the script is adjusted for a different audio source separation task, namely SepiWoogie: https://www.kaggle.com/burekpapi/sepiwoogie